rename server/src/libs/environment/env.ts to serverEnv.ts
