+++ server/src/services/entities/offer.ts
@@ -1,4 +1,4 @@
-import { env } from '../../libs/environment/env'
+import { env } from '../../libs/environment/serverEnv'
 
 import { EntityType, OfferData, TwitterCard } from './types'
